Finishing getting a nix env on macOS
The macOS shell env is now managed by nix and has modern software like
nvim. No more having to edit nix files without a modern editor. yay.

gdb couldn't eval on aarch64-darwin, which is like fine. it's commented
out in dev/common for now.
diffoscope couldn't eval on aarch64-darwin due to a bug in nixpkgs which
fails to keep gnumeric out of the aarch64-darwin inputs like it meant to.
lf- is working on a fix for that. for now it's just commented out in
dev/common as well.