End dependence on brew for pinnable sys utils
Pearl Beta (and likely other machines) had gnu-sed, gnu-grep, and
gnu-which managed by Homebrew formulae. we should likely deprecate
our use of `which` in favor of the `type` shell builtin, but the
other two should become managed by nix.

rust, ruby, node, and perl were also managed by brew formulae, but
compilers and interpreters aren't necessarily as pinnable so they
can be a later issue. These aren't being actively used for the
immediate moment either, so we're uninstalling all remaining brew
formulae and will figure out overlays for these as they're needed. If
brew formulae get used again in the future for just a few compilers and
interpreters, that's probably fine, but we'd like to be making that
decision from scratch later, and for now it feels good to simply tell
brew to uninstall every last formula.