Add eza aliases
Spent a while playing with various ls-alikes, and found that we like
eza. It's got some weird feature gaps and when asked to print
unreasonably much you find it's very not performant and can crash, but
for normal usage it's good, maybe even nice that it hangs for enormous
trees bc it gives us time to interrupt it x3

Part of making it usable for us was to create a bunch of nice little
aliases, so we're adding those here.